disable SSL tests requiring connect to a single instance
