+++ src/test/java/com/splunk/cloudfwd/test/integration/ssl_cert_tests/SslCertValidCloudTrialDisabled.java
@@ -15,42 +15,49 @@ package com.splunk.cloudfwd.test.integration.ssl_cert_tests;/*
  */
 
 import com.splunk.cloudfwd.ConnectionSettings;
+import com.splunk.cloudfwd.HecHealth;
 import com.splunk.cloudfwd.error.HecConnectionTimeoutException;
+
 import com.splunk.cloudfwd.test.util.AbstractConnectionTest;
-import com.splunk.cloudfwd.test.util.BasicCallbacks;
+import org.junit.Assert;
 import org.junit.Test;
 
-import java.util.concurrent.TimeUnit;
+import java.util.List;
 
 /**
- * Cloud>Trial is issued by a private Splunk certificate authority. For 
- * security compliance we should fail it without additional configuration 
- * provided. 
+ * This test enables SSL Verification and attempts to instantiate connection
+ * to a splunk>cloud trial single instance with a splunk>cloud issued SSL cert.
+ * Cloudfwd should accept these certs in cloud.splunk.com domain with a public 
+ * ssl key provided in default cloudfwd.properties file.  
  * 
  * @author ssergeev
  */
-public class SslCertValidCloudTrialDisabledCertValidationIT extends AbstractConnectionTest {
+public class SslCertValidCloudTrialDisabled extends AbstractConnectionTest {
   
-  @Test
-  /**
-   * This test makes sure that connection creation doesn't throw an exception if Cert Validation is disabled 
+  /* 
+   * setUp method in the base class instantiates connection, which blocks and 
+   * awaits for preflight to complete. Having at least one healthy channel 
+   * confirms that SSL connection was properly set up. 
    */
-  public void sendEventSuccessfully() throws InterruptedException, HecConnectionTimeoutException {
-    //no-op
+  @Test
+  public void getHealthCheckPass() throws InterruptedException, HecConnectionTimeoutException {
+    List<HecHealth> healths = super.healthCheck();
+    int healthyCount = (int) healths.stream().filter(e->e.isHealthy()).count();
+    Assert.assertTrue("Expected to get at least one healthy channels, but got "
+            + healthyCount + ", health checks: " + healths, healthyCount > 0);
   }
   
   @Override
   protected void configureProps(ConnectionSettings settings) {
     settings.setUrls("https://input-prd-p-kzgcxv8qsv24.cloud.splunk.com:8088");
     settings.setToken("19FD13FC-8C67-4E5C-8C2B-E39E6CC76152");
-    settings.disableCertValidation();
+    settings.enableCertValidation();
     settings.setMockHttp(false);
-    settings.setSSLCertContent("");
   }
   
   @Override
   protected int getNumEventsToSend() {
     return 0;
   }
-  
+
 }
