+++ src/test/java/com/splunk/cloudfwd/test/integration/ssl_cert_tests/SslCertValidCloudTrialFailByDefaultDisabled.java
@@ -34,7 +34,7 @@ import java.util.concurrent.TimeUnit;
  * 
  * @author ssergeev
  */
-public class SslCertValidCloudTrialFailByDefaultIT extends AbstractConnectionTest {
+public class SslCertValidCloudTrialFailByDefaultDisabled extends AbstractConnectionTest {
 
 
   @Test
